\docType{package}
\name{sweetnR-package}
\alias{sweetnR}
\alias{sweetnR-package}
\title{Syntactic sweetness for R}
\description{
  Convenience functions that simplify coding and
  maintenance
}
\details{
  \tabular{ll}{ Package: \tab sweetnR\cr Type: \tab
  Package\cr Version: \tab 0.0-2\cr Date: \tab
  2013-07-09\cr License: \tab GPL-3\cr }
}
\examples{
# ~~ simple examples of the most important functions ~~
}
\author{
  Kirill Müller <kirill.mueller@ivt.baug.ethz.ch>

  Submit bug reports to:
  \url{https://github.com/krlmlr/sweetnR/issues}
}
\keyword{coding}
\keyword{sugar,}
\keyword{syntax,}

